requirements.txt
mpmath==1.2.1
Pillow==9.1.1
